+++ AWSS3/AWSS3Model.m
@@ -1295,6 +1295,9 @@ return [date aws_stringValue:AWSDateISO8601DateFormat1];
         if ([value caseInsensitiveCompare:@"us-west-2"] == NSOrderedSame) {
             return @(AWSS3BucketLocationConstraintUSWest2);
         }
+        if ([value caseInsensitiveCompare:@"mx-central-1"] == NSOrderedSame) {
+            return @(AWSS3BucketLocationConstraintMXCentral1);
+        }
         return @(AWSS3BucketLocationConstraintUnknown);
     } reverseBlock:^NSString *(NSNumber *value) {
         switch ([value integerValue]) {
@@ -1368,6 +1371,8 @@ return [date aws_stringValue:AWSDateISO8601DateFormat1];
                 return @"us-west-1";
             case AWSS3BucketLocationConstraintUSWest2:
                 return @"us-west-2";
+            case AWSS3BucketLocationConstraintMXCentral1:
+                return @"mx-central-1";
             default:
                 return nil;
         }
@@ -2918,6 +2923,9 @@ return [date aws_stringValue:AWSDateISO8601DateFormat1];
         if ([value caseInsensitiveCompare:@"us-west-2"] == NSOrderedSame) {
             return @(AWSS3BucketLocationConstraintUSWest2);
         }
+        if ([value caseInsensitiveCompare:@"mx-central-1"] == NSOrderedSame) {
+            return @(AWSS3BucketLocationConstraintMXCentral1);
+        }
         return @(AWSS3BucketLocationConstraintUnknown);
     } reverseBlock:^NSString *(NSNumber *value) {
         switch ([value integerValue]) {
@@ -2991,6 +2999,8 @@ return [date aws_stringValue:AWSDateISO8601DateFormat1];
                 return @"us-west-1";
             case AWSS3BucketLocationConstraintUSWest2:
                 return @"us-west-2";
+            case AWSS3BucketLocationConstraintMXCentral1:
+                return @"mx-central-1";
             default:
                 return nil;
         }