+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
--Features for next release
+### New features
+- **AWSCore**
+  - Support for `mx-central-1` - Mexico (see [AWS Regional Services List](https://aws.amazon.com/about-aws/global-infrastructure/regional-product-services/) for a list of services supported in the region)
 
 ## 2.37.2